+++ src/component/FrameContainer/index.tsx
@@ -1,0 +1,12 @@
+import React from "react"
+import { Box } from "@chakra-ui/react"
+
+interface FrameContainerProps {
+    children: React.ReactNode
+}
+
+function FrameContainer({ children }: FrameContainerProps) {
+    return <Box mt={4}>{children}</Box>
+}
+
+export default FrameContainer
