+++ src/page/Home/index.tsx
@@ -1,50 +1,47 @@
-import { Tab, TabList, TabPanel, TabPanels, Tabs } from "@chakra-ui/react"
-import { useEffect, useState } from "react"
+import { HStack, Heading, Stack, Text } from "@chakra-ui/react"
+import { ExternalLink } from "component/ExternalLink"
+import { ExternalLinkIcon } from "@chakra-ui/icons"
 
-import { Connection } from "container/connection"
-import GetStarted from "./component/GetStarted"
-import Position from "./component/Position"
-import TradeComponent from "./component/Trade"
-import LiquidityProvider from "./component/LiquidityProvider"
+import FrameContainer from "component/FrameContainer"
 
-const Home = () => {
-    // NOTE: Focus the trade tab once wallet is connected.
-    const [tabIndex, setTabIndex] = useState(0)
-    const { account } = Connection.useContainer()
-    useEffect(() => {
-        if (account) {
-            setTabIndex(1)
-        } else {
-            setTabIndex(0)
-        }
-    }, [account])
-
+function Home() {
     return (
-        <>
-            {/* Tabs */}
-            <Tabs size="md" mt={5} index={tabIndex} onChange={index => setTabIndex(index)} variant="enclosed">
-                <TabList>
-                    <Tab>Get Started</Tab>
-                    <Tab>Trade</Tab>
-                    <Tab>Positions</Tab>
-                    <Tab>Liquidity Provider</Tab>
-                </TabList>
-                <TabPanels>
-                    <TabPanel py={8} px={3}>
-                        <GetStarted />
-                    </TabPanel>
-                    <TabPanel py={8} px={3}>
-                        <TradeComponent />
-                    </TabPanel>
-                    <TabPanel py={8} px={3}>
-                        <Position />
-                    </TabPanel>
-                    <TabPanel py={8} px={3}>
-                        <LiquidityProvider />
-                    </TabPanel>
-                </TabPanels>
-            </Tabs>
-        </>
+        <FrameContainer>
+            <Stack spacing={10}>
+                <Stack spacing={5}>
+                    <Heading size="md">About</Heading>
+                    <Text>
+                        This is a{" "}
+                        <ExternalLink color="blue.500" href="https://perpdex.com" isExternal>
+                            PerpDEX Protocol <ExternalLinkIcon mx="2px" />
+                        </ExternalLink>{" "}
+                        UI with essential features. Note that you need to manually switch to{" "}
+                        <strong>rinkeby network</strong> to trade.
+                    </Text>
+                    <HStack>
+                        <ExternalLink color="blue.500" href="https://github.com/perpdex/perpdex-lite" isExternal>
+                            Github <ExternalLinkIcon mx="2px" />
+                        </ExternalLink>{" "}
+                        <ExternalLink color="blue.500" href="https://discord.gg/x4cnMGUfyg" isExternal>
+                            Discord <ExternalLinkIcon mx="2px" />
+                        </ExternalLink>
+                    </HStack>
+                </Stack>
+                {/*<Divider />*/}
+                {/*<Stack spacing={5}>*/}
+                {/*    <Heading size="md">Switch Network</Heading>*/}
+                {/*    <Text>*/}
+                {/*        You're currently on <strong>Ethereum Mainnet</strong>{" "}*/}
+                {/*    </Text>*/}
+                {/*    <Button isFullWidth isDisabled colorScheme="blue">*/}
+                {/*        Add / Switch to xDai Network*/}
+                {/*    </Button>*/}
+                {/*    <Text fontSize="sm" color="gray.500">*/}
+                {/*        This feature is not available for now.*/}
+                {/*    </Text>*/}
+                {/*</Stack>*/}
+            </Stack>
+        </FrameContainer>
     )
 }
