+++ src/page/LiquidityProvider/PoolInfo.tsx
@@ -1,6 +1,6 @@
 import { QuestionOutlineIcon } from "@chakra-ui/icons"
 import { Box, Heading, Popover, PopoverBody, PopoverContent, PopoverTrigger, SimpleGrid, Text } from "@chakra-ui/react"
-import { Amm } from "../../../../container/amm"
+import { Amm } from "../../container/amm"
 
 function PoolInfoTable() {
     const { selectedAmm } = Amm.useContainer()
