+++ src/page/LiquidityProvider/index.tsx
@@ -6,10 +6,11 @@ import TitleBar from "./TitleBar"
 import ProvidedInfo from "./ProvidedInfo"
 import Position from "./Position"
 import PoolInfo from "./PoolInfo"
+import FrameContainer from "component/FrameContainer"
 
 function LiquidityProvider() {
     return (
-        <>
+        <FrameContainer>
             <TitleBar />
             <SimpleGrid columns={2} spacing={8} mt="6">
                 <Box borderStyle="solid" borderWidth="1px" borderRadius="12px" p="4">
@@ -25,7 +26,7 @@ function LiquidityProvider() {
                     </VStack>
                 </Box>
             </SimpleGrid>
-        </>
+        </FrameContainer>
     )
 }
